Reorganize Barsukas routes into review and sync packages
